requirements.txt
undetected-chromedriver==3.4.6
selenium==4.11.2
pandas==2.1.1
python-telegram-bot==20.6
